+++ src/library_html5.js
@@ -2669,52 +2669,26 @@ var LibraryJSEvents = {
     return requestAnimationFrame(tick);
   },
 
-  $polyfillSetImmediate__postset:
-    'var __setImmediate_id_counter = 0;\n' +
-    'var __setImmediate_queue = [];\n' +
-    'var __setImmediate_message_id = "_si";\n' +
-    'function __setImmediate_cb(e) {\n' +
-      'if (e.data === __setImmediate_message_id) {\n' +
-        'e.stopPropagation();\n' +
-        '__setImmediate_queue.shift()();\n' +
-        '++__setImmediate_id_counter;\n' +
-      '}\n' +
-    '}\n' +
-    'if (typeof setImmediate === "undefined" && typeof addEventListener === "function") {\n' +
-      'addEventListener("message", __setImmediate_cb, true);\n' +
-      'setImmediate = function(func) {\n' +
-        'postMessage(__setImmediate_message_id, "*");\n' +
-        'return __setImmediate_id_counter + __setImmediate_queue.push(func) - 1;\n' +
-      '}\n' +
-      'clearImmediate = /**@type{function(number=)}*/(function(id) {\n' +
-        'var index = id - __setImmediate_id_counter;\n' +
-        'if (index >= 0 && index < __setImmediate_queue.length) __setImmediate_queue[index] = function(){};\n' + // must preserve the order and count of elements in the queue, so replace the pending callback with an empty function
-      '})\n' +
-    '}',
+  // As setImmediate is only useful in IE11, emscripten has removed support for
+  // it. It is "polyfilled" using a setTimeout.
 
-  $polyfillSetImmediate: function() { /* nop, used for its postset to ensure setImmediate() polyfill is not duplicated between emscripten_set_immediate() and emscripten_set_immediate_loop() if application links to both of them.*/ },
-
-  emscripten_set_immediate__deps: ['$polyfillSetImmediate'],
   emscripten_set_immediate: function(cb, userData) {
-    polyfillSetImmediate();
-    return setImmediate(function() {
+    return setTimeout(function() {
       {{{ makeDynCall('vi', 'cb') }}}(userData);
-    });
+    }, 0);
   },
 
   emscripten_clear_immediate: function(id) {
     clearImmediate(id);
   },
 
-  emscripten_set_immediate_loop__deps: ['$polyfillSetImmediate'],
   emscripten_set_immediate_loop: function(cb, userData) {
-    polyfillSetImmediate();
     function tick() {
       if ({{{ makeDynCall('ii', 'cb') }}}(userData)) {
-        setImmediate(tick);
+        setTimeout(tick, 0);
       }
     }
-    return setImmediate(tick);
+    return setTimeout(tick, 0);
   },
 
   emscripten_set_timeout: function(cb, msecs, userData) {