+++ src/library_pthread.js
@@ -379,8 +379,6 @@ var LibraryPThread = {
           PThread.returnWorkerToPool(worker);
         } else if (cmd === 'objectTransfer') {
           PThread.receiveObjectTransfer(e.data);
-        } else if (e.data.target === 'setimmediate') {
-          worker.postMessage(e.data); // Worker wants to postMessage() to itself to implement setImmediate() emulation.
         } else {
           err("worker sent an unknown command " + cmd);
         }