+++ src/proxyWorker.js
@@ -493,10 +493,6 @@ function onMessageFromMainEmscriptenThread(message) {
       }
       break;
     }
-    case 'setimmediate': {
-      if (Module['setImmediates']) Module['setImmediates'].shift()();
-      break;
-    }
     default: throw 'wha? ' + message.data.target;
   }
 };