+++ tools/cache.py
@@ -78,6 +78,7 @@ class Cache(object):
     shared.safe_ensure_dirs(self.dirname)
 
   def erase(self):
+    return
     self.acquire_cache_lock()
     try:
       if os.path.exists(self.root_dirname):